added print fkt, changed type to float

+++ Regularisierung/Kernel.cl
@@ -1,6 +1,6 @@
 #pragma OPENCL EXTENSION cl_khr_fp64 : enable
 
-__kernel void vecades(__global int *a, __global float *b,  __global int *c, const unsigned int n)
+__kernel void vecades(__global float *a, __global float *b,  __global float *c, const unsigned int n)
 {
 	int idx = get_global_id(0);
